+++ examples/djpcms-site/sitedjpcms/urls.py
@@ -1,6 +1,4 @@
 from djpcms.urls import site_urls
-
-
 urlpatterns = site_urls.patterns()